max_zoom bug on s3 and file caches #311
- added test cases for max_zoom
- removed package stutter s3cache -> s3
- removed package stutter filecache -> file
- updated test cases to use new convetion
